+++ tailcall-typedefs/src/entity.rs
@@ -1,0 +1,7 @@
+// Defines various GraphQL entities used in schema generation.
+#[derive(Clone, Copy)]
+pub enum Entity {
+    Schema,
+    Object,
+    FieldDefinition,
+}
